requirements.txt
streamlit
PyPDF2
transformers
torch
requests
nltk
python-dotenv